+++ packages/core/CHANGELOG.md
@@ -1,6 +1,6 @@
 # @modern-admin/core
 
-## 2.0.0
+## 1.1.0
 
 ### Minor Changes