+++ packages/i18n/CHANGELOG.md
@@ -1,6 +1,6 @@
 # @modern-admin/i18n
 
-## 2.0.0
+## 1.1.0
 
 ### Patch Changes