+++ packages/queue/CHANGELOG.md
@@ -1,6 +1,6 @@
 # @modern-admin/queue
 
-## 2.0.0
+## 1.1.0
 
 ### Patch Changes