+++ packages/ui/CHANGELOG.md
@@ -1,6 +1,6 @@
 # @modern-admin/ui
 
-## 2.0.0
+## 1.1.0
 
 ### Patch Changes